+++ lib/sync_queue.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'src/models/sync_conflict_resolution.dart';
 export 'src/models/sync_entity_ref.dart';
 export 'src/models/sync_entity_state.dart';
 export 'src/models/sync_failure.dart';